Ported all the image processing stuff to OpenCV matrices
git-svn-id: https://svn.salilab.org/imp/trunk@8519 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/em2d/include/PolarResamplingParameters.h
@@ -54,7 +54,7 @@ public:
     matrix_rows_ = rows;
     matrix_cols_ = cols;
     ending_radius_=std::min(rows/2.,cols/2.); // maximum radius, half the size
-    n_rings_ = cv::getOptimalDFTSize(ending_radius_);
+    n_rings_ = cv::getOptimalDFTSize((int)ending_radius_);
     radius_step_ = (ending_radius_-starting_radius_)/((double)(n_rings_));
     parameters_set_ = true;
     IMP_LOG(IMP::VERBOSE,
